Refactor demo app packages
