+++ demo/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.verygoodsecurity">
+    package="com.verygoodsecurity.demo.checkout">
 
     <application
         android:icon="@mipmap/ic_launcher"
@@ -13,7 +13,7 @@
         tools:targetApi="n">
 
         <activity
-            android:name="com.verygoodsecurity.custom.example.MainActivity"
+            android:name="com.verygoodsecurity.demo.checkout.custom.MainActivity"
 
             android:exported="true"
             android:label="@string/app_name_custom_add_cards">
@@ -25,7 +25,7 @@
         </activity>
 
         <activity
-            android:name="com.verygoodsecurity.add.example.AddCardActivity"
+            android:name="com.verygoodsecurity.demo.checkout.add.AddCardActivity"
 
             android:exported="true"
             android:label="@string/app_name_add_cards">
@@ -35,6 +35,5 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
     </application>
 </manifest>
