+++ demo/src/main/java/com/verygoodsecurity/demo/checkout/custom/MainActivity.kt
@@ -1,11 +1,11 @@
-package com.verygoodsecurity.custom.example
+package com.verygoodsecurity.demo.checkout.custom
 
 import android.os.Bundle
 import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.button.MaterialButton
-import com.verygoodsecurity.BuildConfig.STORAGE_ID
-import com.verygoodsecurity.R
+import com.verygoodsecurity.demo.checkout.BuildConfig.STORAGE_ID
+import com.verygoodsecurity.demo.checkout.R
 import com.verygoodsecurity.vgscheckout.VGSCheckout
 import com.verygoodsecurity.vgscheckout.VGSCheckoutCallback
 import com.verygoodsecurity.vgscheckout.config.VGSCheckoutCustomConfig
