+++ demo/src/main/res/drawable/ic_check_circle_black_24dp.xml
