+++ demo/src/main/res/drawable/ic_hot_dog.xml
