+++ demo/src/main/res/drawable/ic_pizza.xml
