+++ demo/src/main/res/drawable/ic_warning_black_24dp.xml
