+++ demo/src/main/res/layout/dialog_payment_transaction_state.xml
