+++ demo/src/main/res/values/colors.xml
