+++ demo/src/main/res/values/dimens.xml
