+++ demo/src/main/res/values/strings.xml
