+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = "VGS Checkout Demo"
-include ':add-cards-example'
+include ':demo'
 include ':vgscheckout'
